+++ Judgement/JudgementRun.cs
@@ -7,11 +7,7 @@ namespace Judgement
 {
     public class JudgementRun : InfiniteTowerRun
     {
-        public int currentWave = 0;
-        public bool shouldGoBazaar = true;
-        public bool isFirstStage = true;
         public Vector3 safeWardPos = Vector3.zero;
-        public Xoroshiro128Plus bazaarRng;
         [SyncVar]
         public Dictionary<NetworkInstanceId, float> persistentHP = new Dictionary<NetworkInstanceId, float>();
         [SyncVar]